+++ app/components/ahoy_captain/stats/container_component.rb
@@ -1,10 +1,11 @@
 module AhoyCaptain
   module Stats
     class ContainerComponent < ViewComponent::Base
-      def initialize(url, label, value)
+      def initialize(url, label, value, selected = false)
         @url = url
         @label = label
         @value = value
+        @selected = selected
       end
     end
   end